Split the deps into api and runtime
This is based on !34 in jfaas. The plan is to only depend on the api in
the pom file and for the runtime to be embedded into the run image.
Currently, there isn't a registry to push a Java image with the runtime
embedded so for run, we'll still reference both the api and runtime in
the pom.