Bug: Fix building of Java project with no deps
Currently there is no way to perform a Docker `COPY` where the source
directory has no files, i.e. no way to ignore the error. So this change
puts all the dependencies into the /function/target directory whereas
they were in the /function/target/dependency directory before. This has
the benefit of knowing that the COPY command will always succeed even if
there are no dependencies as atleast the function jar itself will be in
/function/target directory.