squash# This is a combination of 10 commits2
fn: get available memory related changes

*) Minor getAvailableMemory() related changes.
*) early fail if requested memory too large to meet
*) tracking async and sync pools individually. Sync pool
is reserved for sync jobs only, while async pool can be
used by all jobs.
*) head room estimation for available memory in Linux.
*) for async, dequeue and wait for precise resource
until it is available. This is a naive approach but
it gets rid of extra tuning/options. In other words,
if we dequeue an async msg, we assume at some
point the server should be able to service it.