add a license printing script
this takes info from glide.yaml, which covers all our external dependencies,
and then searches for a LICENSE in each of them. the good news is that it
covers all of our stuff right now, there are none without a license we can't
find. the bad news is, it's not very robust. this seemed to be a manual
process so this is markedly better.

removed one dep we had in glide.yaml that we aren't actually using, should go
away next update